split up filc main.go and make flags variables instead of functions
