+++ filc/main.go
@@ -3,30 +3,10 @@ package main
 import (
 	"fmt"
 	"os"
-	"strings"
-	"time"
 
-	"github.com/application-research/filclient"
-	"github.com/application-research/filclient/retrievehelper"
-	"github.com/dustin/go-humanize"
-	"github.com/filecoin-project/go-address"
-	cborutil "github.com/filecoin-project/go-cbor-util"
-	datatransfer "github.com/filecoin-project/go-data-transfer"
-	"github.com/filecoin-project/go-fil-markets/retrievalmarket"
-	"github.com/filecoin-project/go-fil-markets/storagemarket"
-	"github.com/filecoin-project/go-state-types/big"
-	"github.com/filecoin-project/lotus/chain/types"
-	lcli "github.com/filecoin-project/lotus/cli"
-	"github.com/ipfs/go-blockservice"
-	"github.com/ipfs/go-cid"
-	chunker "github.com/ipfs/go-ipfs-chunker"
-	offline "github.com/ipfs/go-ipfs-exchange-offline"
 	logging "github.com/ipfs/go-log"
-	"github.com/ipfs/go-merkledag"
-	"github.com/ipfs/go-unixfs/importer"
 	"github.com/mitchellh/go-homedir"
 	cli "github.com/urfave/cli/v2"
-	"golang.org/x/xerrors"
 )
 
 func main() {
@@ -83,582 +63,5 @@ func ddir(cctx *cli.Context) string {
 		return ddir
 	default:
 		panic("ddir should be present in CLI metadata")
-	}
-}
-
-// Read a single miner from the CLI, erroring if more than one is specified, or
-// none are present.
-func readMiner(cctx *cli.Context) (address.Address, error) {
-	miners, err := readMiners(cctx)
-	if err != nil {
-		return address.Undef, err
-	}
-
-	if len(miners) > 1 {
-		return address.Undef, fmt.Errorf("only one miner expected")
-	}
-
-	return miners[0], nil
-}
-
-// Read a comma-separated list of miners from the CLI, erroring if none are
-// present.
-func readMiners(cctx *cli.Context) ([]address.Address, error) {
-	minerStrings := strings.Split(cctx.String("miner"), ",")
-
-	if len(minerStrings) == 0 {
-		return nil, fmt.Errorf("no miners were specified")
-	}
-
-	miners := make([]address.Address, len(minerStrings))
-	for i, ms := range minerStrings {
-		miner, err := address.NewFromString(ms)
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse miner %s: %w", ms, err)
-		}
-
-		miners[i] = miner
-	}
-
-	return miners, nil
-}
-
-func flagMiner() *cli.StringFlag {
-	flag := &cli.StringFlag{
-		Name:    "miner",
-		Aliases: []string{"miners", "m"},
-	}
-
-	return flag
-}
-
-func flagMinerRequired() *cli.StringFlag {
-	flag := flagMiner()
-	flag.Required = true
-
-	return flag
-}
-
-func flagVerified() *cli.BoolFlag {
-	flag := &cli.BoolFlag{
-		Name: "verified",
-	}
-
-	return flag
-}
-
-func flagOutput() *cli.PathFlag {
-	flag := &cli.PathFlag{
-		Name:    "output",
-		Aliases: []string{"o"},
-	}
-
-	return flag
-}
-
-var makeDealCmd = &cli.Command{
-	Name:      "deal",
-	Usage:     "Make a storage deal with a miner",
-	ArgsUsage: "<file path>",
-	Flags: []cli.Flag{
-		flagMinerRequired(),
-		flagVerified(),
-	},
-	Action: func(cctx *cli.Context) error {
-		if !cctx.Args().Present() {
-			return fmt.Errorf("please specify file to make deal for")
-		}
-
-		ddir := ddir(cctx)
-
-		mstr := cctx.String("miner")
-		if mstr == "" {
-			return fmt.Errorf("must specify miner to make deals with")
-		}
-
-		miner, err := address.NewFromString(mstr)
-		if err != nil {
-			return err
-		}
-
-		nd, err := setup(cctx.Context, ddir)
-		if err != nil {
-			return err
-		}
-
-		fc, closer, err := clientFromNode(cctx, nd, ddir)
-		if err != nil {
-			return err
-		}
-		defer closer()
-
-		fi, err := os.Open(cctx.Args().First())
-		if err != nil {
-			return err
-		}
-
-		tpr := func(s string, args ...interface{}) {
-			fmt.Printf("[%s] "+s+"\n", append([]interface{}{time.Now().Format("15:04:05")}, args...)...)
-		}
-
-		bserv := blockservice.New(nd.Blockstore, nil)
-		dserv := merkledag.NewDAGService(bserv)
-
-		tpr("importing file...")
-		spl := chunker.DefaultSplitter(fi)
-
-		obj, err := importer.BuildDagFromReader(dserv, spl)
-		if err != nil {
-			return err
-		}
-
-		tpr("File CID: %s", obj.Cid())
-
-		ask, err := fc.GetAsk(cctx.Context, miner)
-		if err != nil {
-			return err
-		}
-
-		verified := cctx.Bool("verified")
-
-		price := ask.Ask.Ask.Price
-		if verified {
-			price = ask.Ask.Ask.VerifiedPrice
-		}
-
-		proposal, err := fc.MakeDeal(cctx.Context, miner, obj.Cid(), price, 0, 2880*365, verified)
-		if err != nil {
-			return err
-		}
-
-		propnd, err := cborutil.AsIpld(proposal.DealProposal)
-		if err != nil {
-			return xerrors.Errorf("failed to compute deal proposal ipld node: %w", err)
-		}
-
-		tpr("proposal cid: %s", propnd.Cid())
-
-		if err := saveDealProposal(ddir, propnd.Cid(), proposal.DealProposal); err != nil {
-			return err
-		}
-
-		resp, err := fc.SendProposal(cctx.Context, proposal)
-		if err != nil {
-			return err
-		}
-
-		tpr("response state: %d", resp.Response.State)
-		switch resp.Response.State {
-		case storagemarket.StorageDealError:
-			return fmt.Errorf("error response from miner: %s", resp.Response.Message)
-		case storagemarket.StorageDealProposalRejected:
-			return fmt.Errorf("deal rejected by miner: %s", resp.Response.Message)
-		default:
-			return fmt.Errorf("unrecognized response from miner: %d %s", resp.Response.State, resp.Response.Message)
-		case storagemarket.StorageDealWaitingForData, storagemarket.StorageDealProposalAccepted:
-			tpr("miner accepted the deal!")
-		}
-
-		tpr("starting data transfer... %s", resp.Response.Proposal)
-
-		chanid, err := fc.StartDataTransfer(cctx.Context, miner, resp.Response.Proposal, obj.Cid())
-		if err != nil {
-			return err
-		}
-
-		var lastStatus datatransfer.Status
-	loop:
-		for {
-			status, err := fc.TransferStatus(cctx.Context, chanid)
-			if err != nil {
-				return err
-			}
-
-			switch status.Status {
-			case datatransfer.Failed:
-				return fmt.Errorf("data transfer failed: %s", status.Message)
-			case datatransfer.Cancelled:
-				return fmt.Errorf("transfer cancelled: %s", status.Message)
-			case datatransfer.Failing:
-				tpr("data transfer failing... %s", status.Message)
-				// I guess we just wait until its failed all the way?
-			case datatransfer.Requested:
-				if lastStatus != status.Status {
-					tpr("data transfer requested")
-				}
-				//fmt.Println("transfer is requested, hasnt started yet")
-				// probably okay
-			case datatransfer.TransferFinished, datatransfer.Finalizing, datatransfer.Completing:
-				if lastStatus != status.Status {
-					tpr("current state: %s", status.StatusStr)
-				}
-			case datatransfer.Completed:
-				tpr("transfer complete!")
-				break loop
-			case datatransfer.Ongoing:
-				fmt.Printf("[%s] transfer progress: %d      \n", time.Now().Format("15:04:05"), status.Sent)
-			default:
-				tpr("Unexpected data transfer state: %d (msg = %s)", status.Status, status.Message)
-			}
-			time.Sleep(time.Millisecond * 100)
-			lastStatus = status.Status
-		}
-
-		tpr("transfer completed, miner: %s, propcid: %s %s", miner, resp.Response.Proposal, propnd.Cid())
-
-		return nil
-	},
-}
-
-var infoCmd = &cli.Command{
-	Name:      "info",
-	Usage:     "Display wallet information",
-	ArgsUsage: " ",
-	Action: func(cctx *cli.Context) error {
-		ddir := ddir(cctx)
-
-		nd, err := setup(cctx.Context, ddir)
-		if err != nil {
-			return err
-		}
-
-		api, closer, err := lcli.GetGatewayAPI(cctx)
-		if err != nil {
-			return err
-		}
-		defer closer()
-
-		addr, err := nd.Wallet.GetDefault()
-		if err != nil {
-			return err
-		}
-
-		balance := big.NewInt(0)
-		verifiedBalance := big.NewInt(0)
-
-		act, err := api.StateGetActor(cctx.Context, addr, types.EmptyTSK)
-		if err != nil {
-			fmt.Println("NOTE - Actor not found on chain")
-		} else {
-			balance = act.Balance
-
-			v, err := api.StateVerifiedClientStatus(cctx.Context, addr, types.EmptyTSK)
-			if err != nil {
-				return err
-			}
-
-			verifiedBalance = *v
-		}
-
-		fmt.Printf("Default client address: %v\n", addr)
-		fmt.Printf("Balance:                %v\n", types.FIL(balance))
-		fmt.Printf("Verified Balance:       %v\n", types.FIL(verifiedBalance))
-
-		return nil
-	},
-}
-
-var getAskCmd = &cli.Command{
-	Name:      "get-ask",
-	Usage:     "Query storage deal ask for a miner",
-	ArgsUsage: "<miner>",
-	Action: func(cctx *cli.Context) error {
-		if !cctx.Args().Present() {
-			return fmt.Errorf("please specify miner to query ask of")
-		}
-
-		ddir := ddir(cctx)
-
-		miner, err := address.NewFromString(cctx.Args().First())
-		if err != nil {
-			return err
-		}
-
-		fc, closer, err := getClient(cctx, ddir)
-		if err != nil {
-			return err
-		}
-		defer closer()
-
-		ask, err := fc.GetAsk(cctx.Context, miner)
-		if err != nil {
-			return fmt.Errorf("failed to get ask: %s", err)
-		}
-
-		printAskResponse(ask.Ask.Ask)
-
-		return nil
-	},
-}
-
-var listDealsCmd = &cli.Command{
-	Name:      "list",
-	Usage:     "List local storage deal history",
-	ArgsUsage: " ",
-	Action: func(cctx *cli.Context) error {
-		ddir := ddir(cctx)
-
-		deals, err := listDeals(ddir)
-		if err != nil {
-			return err
-		}
-
-		for _, dcid := range deals {
-			fmt.Println(dcid)
-		}
-
-		return nil
-	},
-}
-
-var retrieveFileCmd = &cli.Command{
-	Name:        "retrieve",
-	Usage:       "Retrieve a file by CID from a miner",
-	Description: "Retrieve a file by CID from a miner. If desired, multiple miners can be specified as fallbacks in case of a failure (comma-separated, no spaces).",
-	ArgsUsage:   "<cid>",
-	Flags: []cli.Flag{
-		flagMinerRequired(),
-		flagOutput(),
-	},
-	Action: func(cctx *cli.Context) error {
-		cidStr := cctx.Args().First()
-		if cidStr == "" {
-			return fmt.Errorf("please specify a CID to retrieve")
-		}
-
-		miners, err := readMiners(cctx)
-		if err != nil {
-			return err
-		}
-
-		outputStr := cctx.String("output")
-		if outputStr == "" {
-			outputStr = cidStr
-		}
-
-		c, err := cid.Decode(cidStr)
-		if err != nil {
-			return err
-		}
-
-		ddir := ddir(cctx)
-
-		node, err := setup(cctx.Context, ddir)
-		if err != nil {
-			return err
-		}
-
-		fc, closer, err := clientFromNode(cctx, node, ddir)
-		if err != nil {
-			return err
-		}
-		defer closer()
-
-		// Attempt the file retrieval from each miner
-		for _, miner := range miners {
-			fmt.Println("attempting retrieval with miner", miner)
-
-			ask, err := fc.RetrievalQuery(cctx.Context, miner, c)
-			if err != nil {
-				fmt.Println(err)
-				continue
-			}
-
-			proposal, err := retrievehelper.RetrievalProposalForAsk(ask, c, nil)
-			if err != nil {
-				fmt.Println(err)
-				continue
-			}
-
-			stats, err := fc.RetrieveContent(cctx.Context, miner, proposal)
-			if err != nil {
-				fmt.Println(err)
-				continue
-			}
-
-			bserv := blockservice.New(node.Blockstore, offline.Exchange(node.Blockstore))
-			dserv := merkledag.NewDAGService(bserv)
-
-			dnode, err := dserv.Get(cctx.Context, c)
-			if err != nil {
-				fmt.Println(err)
-				continue
-			}
-
-			if err := os.WriteFile(outputStr, dnode.RawData(), 0644); err != nil {
-				fmt.Println(err)
-				continue
-			}
-
-			printRetrievalStats(stats)
-
-			fmt.Println("Saved output to", outputStr)
-
-			// If we completed the full retrieval without any problems, we're
-			// done, so return immediately without trying any other miners
-			return nil
-		}
-
-		// If we fell out of that loop, it means none of the miners worked
-		return fmt.Errorf("retrieval failed for all miners")
-	},
-}
-
-var queryRetrievalCmd = &cli.Command{
-	Name:      "query-retrieval",
-	Usage:     "Query retrieval information for a CID",
-	ArgsUsage: "<cid>",
-	Flags: []cli.Flag{
-		&cli.StringFlag{Name: "miner", Aliases: []string{"m"}, Required: true},
-	},
-	Action: func(cctx *cli.Context) error {
-
-		cidStr := cctx.Args().First()
-		if cidStr == "" {
-			return fmt.Errorf("please specify a CID to query retrieval of")
-		}
-
-		minerStr := cctx.String("miner")
-		if minerStr == "" {
-			return fmt.Errorf("must specify a miner with --miner")
-		}
-
-		cid, err := cid.Decode(cidStr)
-		if err != nil {
-			return err
-		}
-
-		miner, err := address.NewFromString(minerStr)
-		if err != nil {
-			return err
-		}
-
-		ddir := ddir(cctx)
-
-		fc, closer, err := getClient(cctx, ddir)
-		if err != nil {
-			return err
-		}
-		defer closer()
-
-		query, err := fc.RetrievalQuery(cctx.Context, miner, cid)
-		if err != nil {
-			return err
-		}
-
-		printQueryResponse(query)
-
-		return nil
-	},
-}
-
-var clearBlockstoreCmd = &cli.Command{
-	Name:      "clear-blockstore",
-	Usage:     "Delete all retrieved file data in the blockstore",
-	ArgsUsage: " ",
-	Action: func(cctx *cli.Context) error {
-		ddir := ddir(cctx)
-
-		fmt.Println("clearing blockstore...")
-
-		if err := os.RemoveAll(blockstorePath(ddir)); err != nil {
-			return err
-		}
-
-		fmt.Println("done")
-
-		return nil
-	},
-}
-
-func printAskResponse(ask *storagemarket.StorageAsk) {
-	fmt.Printf(`ASK RESPONSE
------
-Miner: %v
-Price (Unverified): %v (%v)
-Price (Verified): %v (%v)
-Min Piece Size: %v
-Max Piece Size: %v
-`,
-		ask.Miner,
-		ask.Price, types.FIL(ask.Price),
-		ask.VerifiedPrice, types.FIL(ask.VerifiedPrice),
-		ask.MinPieceSize,
-		ask.MaxPieceSize,
-	)
-}
-
-func printRetrievalStats(stats *filclient.RetrievalStats) {
-	fmt.Printf(`RETRIEVAL STATS
------
-Size:          %v (%v)
-Total Payment: %v (%v)
-Ask Price:     %v (%v)
-Num Payments:  %v
-Duration:      %v
-Average Speed: %v (%v/s)
-Peer:          %v
-`,
-		stats.Size, humanize.IBytes(stats.Size),
-		stats.TotalPayment, types.FIL(stats.TotalPayment),
-		stats.AskPrice, types.FIL(stats.AskPrice),
-		stats.NumPayments,
-		stats.Duration,
-		stats.AverageSpeed, humanize.IBytes(stats.AverageSpeed),
-		stats.Peer,
-	)
-}
-
-func printQueryResponse(query *retrievalmarket.QueryResponse) {
-	var status string
-	switch query.Status {
-	case retrievalmarket.QueryResponseAvailable:
-		status = "Available"
-	case retrievalmarket.QueryResponseUnavailable:
-		status = "Unavailable"
-	case retrievalmarket.QueryResponseError:
-		status = "Error"
-	default:
-		status = fmt.Sprintf("Unrecognized Status (%d)", query.Status)
-	}
-
-	var pieceCIDFound string
-	switch query.PieceCIDFound {
-	case retrievalmarket.QueryItemAvailable:
-		pieceCIDFound = "Available"
-	case retrievalmarket.QueryItemUnavailable:
-		pieceCIDFound = "Unavailable"
-	case retrievalmarket.QueryItemUnknown:
-		pieceCIDFound = "Unknown"
-	default:
-		pieceCIDFound = fmt.Sprintf("Unrecognized (%d)", query.PieceCIDFound)
-	}
-
-	total := big.Add(query.UnsealPrice, big.Mul(big.NewIntUnsigned(query.Size), query.MinPricePerByte))
-	fmt.Printf(`QUERY RESPONSE
------
-Status:                        %v
-Piece CID Found:               %v
-Size:                          %v (%v)
-Unseal Price:                  %v (%v)
-Min Price Per Byte:            %v (%v)
-Total Retrieval Price:         %v (%v)
-Payment Address:               %v
-Max Payment Interval:          %v (%v)
-Max Payment Interval Increase: %v (%v)
-`,
-		status,
-		pieceCIDFound,
-		query.Size, humanize.IBytes(query.Size),
-		query.UnsealPrice, types.FIL(query.UnsealPrice),
-		query.MinPricePerByte, types.FIL(query.MinPricePerByte),
-		total, types.FIL(total),
-		query.PaymentAddress,
-		query.MaxPaymentInterval, humanize.IBytes(query.MaxPaymentInterval),
-		query.MaxPaymentIntervalIncrease, humanize.IBytes(query.MaxPaymentIntervalIncrease),
-	)
-
-	if query.Message != "" {
-		fmt.Printf("Message: %v\n", query.Message)
 	}
 }
